+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 ## v0.0.9 (not yet released)
 
 *   Created a structure representing the results and status of a conversion, and reworked the main entry point functions to use it.
+*   Allow specification of the directory for application dependencies (i.e. where to look for include_lib files.)
 *   Simple macros can now appear in typespecs. The transpiler substitutes the replacement value because Elixir syntax can't handle a macro call at that location.
 *   Rename specs when the transpiler renames the associated function.
 *   Reworked conversion of "if" statements so it supports expressions that cause errors.